requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
faiss-cpu
pypdf
python-dotenv
sentence-transformers
pymupdf 
tqdm
numpy
streamlit
PyPDF2
ipykernel
pymupdf
huggingface_hub
langchain_huggingface
google.generativeai